reorganize delta code
