+++ src/react/deltas.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { mergeDeltas, applyDeltasToStreamMessage } from "./deltas.js";
+import { mergeDeltas, applyDeltasToStreamMessage } from "../deltas.js";
 import type { StreamMessage, StreamDelta } from "../validators.js";
 import { omit } from "convex-helpers";
 import type { TextStreamPart, ToolSet } from "ai";
