+++ src/react/useThreadMessages.ts
@@ -12,7 +12,7 @@ import { useMemo, useRef, useState } from "react";
 import type { MessageDoc } from "../client/index.js";
 import type { SyncStreamsReturnValue } from "../client/types.js";
 import type { StreamArgs } from "../validators.js";
-import { mergeDeltas } from "./deltas.js";
+import { mergeDeltas } from "../deltas.js";
 import type {
   ThreadQuery,
   ThreadStreamQuery,
